refect : login-user merge
