--- backend/src/main/java/com/example/backend/login/model/repository/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.backend.login.model.repository;
-
-import com.example.backend.login.model.entity.User;
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findBySocialId(String socialId);
-}
