+++ backend/src/main/java/com/example/backend/user/model/repository/UserRepository.java
@@ -7,4 +7,7 @@ import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
+    Optional<User> findBySocialId(String socialId);
+
+
 }
